Feature/edit delete

+++ client/src/utilities/category/category-services.ts
@@ -1,4 +1,5 @@
 import * as categoriesApi from './category-api';
+import * as tools from '../tools';
 import { Category } from '../types';
 
 export async function getAllCategories() {
@@ -12,10 +13,10 @@ export async function getAllCategories() {
 }
 
 export async function getCategory(id: string | undefined) {
-    if(!id){
-        return 'Error: id not defined.';
-    }
     try {
+        if(!id){
+            throw Error('Error: id undefined.');
+        }
         const res = await categoriesApi.show(id);
         return res.data;
     } catch (err) {
@@ -25,15 +26,19 @@ export async function getCategory(id: string | undefined) {
 
 export async function createCategory(data: Category) {
     try {
-        const res = await categoriesApi.create(data);
+        const title: string = tools.capitalize(data.title);
+        const res = await categoriesApi.create({...data, title: title});
         return res.data;
     } catch (err) {
         return err;
     }
 }
 
-export async function updateCategory(id: string, data: Category) {
+export async function updateCategory(id: string | undefined, data: Category) {
     try {
+        if(!id){
+            throw Error('Error: id undefined.');
+        }
         const res = await categoriesApi.update(id, data);
         return res.data;
     } catch (err) {
@@ -41,8 +46,11 @@ export async function updateCategory(id: string, data: Category) {
     }
 }
 
-export async function destroyCategory(id: string) {
+export async function destroyCategory(id: string | undefined) {
     try {
+        if(!id){
+            throw Error('Error: id undefined.');
+        }
         const res = await categoriesApi.destroy(id);
         return res.data;
     } catch (err) {
