+++ client/src/utilities/submission/submission-services.ts
@@ -1,5 +1,7 @@
 import * as submissionsApi from './submission-api';
-import { Submission } from '../types';
+import * as userServices from '../user/user-services';
+import * as localStorageTools from '../local-storage';
+import { Submission, User } from '../types';
 
 export async function getAllSubmissions() {
     try {
@@ -10,8 +12,11 @@ export async function getAllSubmissions() {
     }
 }
 
-export async function getSubmission(id: string) {
+export async function getSubmission(id: string | undefined) {
     try {
+        if (!id) {
+            throw Error('Error: id undefined.');
+        }
         const res = await submissionsApi.show(id);
         return res.data;
     } catch (err) {
@@ -19,14 +24,29 @@ export async function getSubmission(id: string) {
     }
 }
 
-export async function getSubmissionList(ids: Array<string>): Promise<any> {
+export async function getSubmissionList(user: User): Promise<any> {
     try {
         const data: Array<Submission> = [];
-        for (let id of ids){
-            const res = await submissionsApi.show(id);
-            data.push(res.data);
+        if (user.submissions.length) {
+            for (let i = 0; i < user.submissions.length; i++) {
+                await submissionsApi.show(user.submissions[i]).then(async (res) => {
+                    if (res.data) {
+                        data.push(res.data);
+                    } else {
+                        const idsCache: Array<string> = [...user.submissions];
+                        idsCache.splice(i, 1);
+                        await userServices.addSubmission(user.id, idsCache).then((updatedUser: User)=>{
+                            const currentUser = localStorageTools.getUser();
+                            if(user.id === currentUser.id){
+                                localStorageTools.setUser(updatedUser);
+                            }
+                        });
+                    }
+                });
+            }
         }
         return data;
+
     } catch (err) {
         return err;
     }
@@ -41,8 +61,11 @@ export async function createSubmission(data: Submission) {
     }
 }
 
-export async function updateSubmission(id: string, data: Submission) {
+export async function updateSubmission(id: string | undefined, data: Submission) {
     try {
+        if (!id) {
+            throw Error('Error: id undefined.');
+        }
         const res = await submissionsApi.update(id, data);
         return res.data;
     } catch (err) {
@@ -50,8 +73,11 @@ export async function updateSubmission(id: string, data: Submission) {
     }
 }
 
-export async function destroySubmission(id: string) {
+export async function destroySubmission(id: string | undefined) {
     try {
+        if (!id) {
+            throw Error('Error: id undefined.');
+        }
         const res = await submissionsApi.destroy(id);
         return res.data;
     } catch (err) {
